# geo-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GeoLocation API",
        "title": "GeoLocation API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Mahmoud AllamehAmiri",
            "email": "m.allamehamiri@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/v1/geo-location/{ip}": {
            "get": {
                "description": "This service will try to find geolocation of the given IP address",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Geo"
                ],
                "summary": "Get Geolocation Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": " ",
                        "name": "ip",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GeoLocationResponse",
                        "schema": {
                            "$ref": "#/definitions/api.GeoLocationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.AppError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.AppError": {
            "type": "object",
            "properties": {
                "error_data": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "errorParam": "value"
                    }
                },
                "error_message": {
                    "type": "string"
                }
            }
        },
        "api.GeoLocationResponse": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "country_code": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "mystery_value": {
                    "type": "string"
                }
            }
        }
    }
}